// src/tinymce.d.ts
interface AlertBannerSpec {
  type: 'alertbanner';
  level: 'info' | 'warn' | 'error' | 'success';
  text: string;
  icon: string;
  url?: string;
}
interface ButtonSpec {
  type: 'button';
  text: string;
  enabled?: boolean;
  primary?: boolean;
  name?: string;
  icon?: string;
  borderless?: boolean;
  buttonType?: 'primary' | 'secondary' | 'toolbar';
}
interface FormComponentSpec {
  type: string;
  name: string;
}
interface FormComponentWithLabelSpec extends FormComponentSpec {
  label?: string;
}
interface CheckboxSpec extends FormComponentSpec {
  type: 'checkbox';
  label: string;
  enabled?: boolean;
}
interface CollectionSpec extends FormComponentWithLabelSpec {
  type: 'collection';
}
interface CollectionItem {
  value: string;
  text: string;
  icon: string;
}
interface ColorInputSpec extends FormComponentWithLabelSpec {
  type: 'colorinput';
  storageKey?: string;
}
interface ColorPickerSpec extends FormComponentWithLabelSpec {
  type: 'colorpicker';
}
interface CustomEditorInit {
  setValue: (value: string) => void;
  getValue: () => string;
  destroy: () => void;
}
type CustomEditorInitFn = (elm: HTMLElement, settings: any) => Promise<CustomEditorInit>;
interface CustomEditorOldSpec extends FormComponentSpec {
  type: 'customeditor';
  tag?: string;
  init: (e: HTMLElement) => Promise<CustomEditorInit>;
}
interface CustomEditorNewSpec extends FormComponentSpec {
  type: 'customeditor';
  tag?: string;
  scriptId: string;
  scriptUrl: string;
  settings?: any;
}
type CustomEditorSpec = CustomEditorOldSpec | CustomEditorNewSpec;
interface DropZoneSpec extends FormComponentWithLabelSpec {
  type: 'dropzone';
}
interface GridSpec {
  type: 'grid';
  columns: number;
  items: BodyComponentSpec[];
}
interface HtmlPanelSpec {
  type: 'htmlpanel';
  html: string;
  presets?: 'presentation' | 'document';
}
interface IframeSpec extends FormComponentWithLabelSpec {
  type: 'iframe';
  border?: boolean;
  sandboxed?: boolean;
  streamContent?: boolean;
  transparent?: boolean;
}
interface ImagePreviewSpec extends FormComponentSpec {
  type: 'imagepreview';
  height?: string;
}
interface InputSpec extends FormComponentWithLabelSpec {
  type: 'input';
  inputMode?: string;
  placeholder?: string;
  maximized?: boolean;
  enabled?: boolean;
}
type Alignment = 'start' | 'center' | 'end';
interface LabelSpec {
  type: 'label';
  label: string;
  items: BodyComponentSpec[];
  align?: Alignment;
}
interface ListBoxSingleItemSpec {
  text: string;
  value: string;
}
interface ListBoxNestedItemSpec {
  text: string;
  items: ListBoxItemSpec[];
}
type ListBoxItemSpec = ListBoxNestedItemSpec | ListBoxSingleItemSpec;
interface ListBoxSpec extends FormComponentWithLabelSpec {
  type: 'listbox';
  items: ListBoxItemSpec[];
  disabled?: boolean;
}
interface PanelSpec {
  type: 'panel';
  classes?: string[];
  items: BodyComponentSpec[];
}
interface SelectBoxItemSpec {
  text: string;
  value: string;
}
interface SelectBoxSpec extends FormComponentWithLabelSpec {
  type: 'selectbox';
  items: SelectBoxItemSpec[];
  size?: number;
  enabled?: boolean;
}
interface SizeInputSpec extends FormComponentWithLabelSpec {
  type: 'sizeinput';
  constrain?: boolean;
  enabled?: boolean;
}
interface SliderSpec extends FormComponentSpec {
  type: 'slider';
  label: string;
  min?: number;
  max?: number;
}
interface TableSpec {
  type: 'table';
  header: string[];
  cells: string[][];
}
interface TextAreaSpec extends FormComponentWithLabelSpec {
  type: 'textarea';
  placeholder?: string;
  maximized?: boolean;
  enabled?: boolean;
}
// ...
type Id = string;
interface TreeSpec {
  type: 'tree';
  items: TreeItemSpec[];
  onLeafAction?: (id: Id) => void;
  defaultExpandedIds?: Id[];
  onToggleExpand?: (expandedIds: Id[], { expanded, node }: {
    expanded: boolean;
    node: Id;
  }) => void;
  defaultSelectedId?: Id;
}
interface BaseTreeItemSpec {
  title: string;
  id: Id;
  menu?: ToolbarMenuButtonSpec;
}
interface DirectorySpec extends BaseTreeItemSpec {
  type: 'directory';
  children: TreeItemSpec[];
}
interface LeafSpec extends BaseTreeItemSpec {
  type: 'leaf';
}
type TreeItemSpec = DirectorySpec | LeafSpec;
interface UrlInputSpec extends FormComponentWithLabelSpec {
  type: 'urlinput';
  filetype?: 'image' | 'media' | 'file';
  enabled?: boolean;
  picker_text?: string;
}
interface UrlInputData {
  value: string;
  meta: {
    text?: string;
  };
}
export type BodyComponentSpec = BarSpec | ButtonSpec | CheckboxSpec | TextAreaSpec | InputSpec | ListBoxSpec | SelectBoxSpec | SizeInputSpec | SliderSpec | IframeSpec | HtmlPanelSpec | UrlInputSpec | DropZoneSpec | ColorInputSpec | GridSpec | ColorPickerSpec | ImagePreviewSpec | AlertBannerSpec | CollectionSpec | LabelSpec | TableSpec | TreeSpec | PanelSpec | CustomEditorSpec;
interface BarSpec {
  type: 'bar';
  items: BodyComponentSpec[];
}